requirements.txt
duckdb==1.4.0
numpy==2.3.3
pandas==2.3.2
pyarrow==21.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
tzdata==2025.2
